LayoutTests/js/dom/modules/nomodule-dynamic-classic-src-expected.txt

Test dynamic "src" classic script with nomodule.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS executed is false
PASS successfullyParsed is true

TEST COMPLETE
